% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.r
\name{abstract_topics}
\alias{abstract_topics}
\title{Topic models for abstracts}
\usage{
abstract_topics(wos, topic_count = 20, extra_stopwords = "")
}
\arguments{
\item{wos:}{the dataframe of references}

\item{extra_stopwords:}{any extra stopwords to exclude}

\item{topic_count:}{the numbers of topics; default is 20}
}
\description{
Topic models for abstracts
}
\examples{
abstract_topics()
}
\keyword{abstracts}
