% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.r
\name{abstract_topics_plot}
\alias{abstract_topics_plot}
\title{Plot topic models for abstracts}
\usage{
abstract_topics_plot(tidy_lda)
}
\arguments{
\item{tidy_lda:}{a tidytext lda object}
}
\description{
Plot topic models for abstracts
}
\examples{
abstract_topics_plot()
}
\keyword{abstracts}
