% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.r
\name{authors_top}
\alias{authors_top}
\title{Display top n first authors}
\usage{
authors_top(wos, n = 30)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{n:}{the number of authors to return}
}
\description{
Display top n first authors
}
\examples{
authors_top()
}
\keyword{authors}
