% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{cited_references_gather}
\alias{cited_references_gather}
\title{List cited references in narrow form 
The key for this data frame is title and publication year}
\usage{
cited_references_gather(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
List cited references in narrow form 
The key for this data frame is title and publication year
}
\examples{
cited_references_gather()
}
\keyword{citations}
