% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cited_reference_time.r
\name{cited_references_time}
\alias{cited_references_time}
\title{Plot top cited references over time}
\usage{
cited_references_time(wos, topn = 20)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{n}{the number of cited references to plot}
}
\description{
Plot top cited references over time
}
\examples{
cited_references_time()
}
\keyword{citations}
