% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.r
\name{construct_stopword_list}
\alias{construct_stopword_list}
\title{Add tot he stopword list}
\usage{
construct_stopword_list(extra_stopwords)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{extra_stopwords:}{any extra stopwords to exclude}
}
\description{
Add tot he stopword list
}
\examples{
construct_stopword_list()
}
\keyword{stopwords}
