% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_format.r
\name{convert_file_to_bib}
\alias{convert_file_to_bib}
\title{Convert all WoS isi csv in directory to bibtex a file}
\usage{
convert_file_to_bib(wos_dir = ".", out_file = "out.bib")
}
\arguments{
\item{wos_dir:}{the directory of references}

\item{out_file:}{the bib file produced
convert_file_to_bib()}
}
\description{
Convert all WoS isi csv in directory to bibtex a file
}
\keyword{bibtex}
