% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_format.r
\name{convert_refs_to_bib}
\alias{convert_refs_to_bib}
\title{Convert all WoS records to biblatex and save as file}
\usage{
convert_refs_to_bib(wos, out_file = "out.bib")
}
\arguments{
\item{wos:}{the references}

\item{out_file:}{the bib file produced
convert_refs_to_bib()}
}
\description{
Convert all WoS records to biblatex and save as file
}
\keyword{bibtex}
