% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discplines.r
\name{disciplines}
\alias{disciplines}
\title{discplines from WoS WC field separated}
\usage{
disciplines(wos, with_id = FALSE)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
The function filters only for journal articles
}
\examples{
disciplines()
}
\keyword{discipline}
