% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.r
\name{field_tfidf}
\alias{field_tfidf}
\title{Count term frequency/inverse document frequencies}
\usage{
field_tfidf(wos, field = "AB", extra_stopwords = "")
}
\arguments{
\item{wos:}{the dataframe of references}

\item{field:}{field to construct tfidf matrix for}

\item{extra_stopwords:}{any extra stopwords to exclude}
}
\description{
Count term frequency/inverse document frequencies
}
\examples{
field_tfidf()
}
\keyword{tfidf}
