% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.r
\name{keyword_count_top}
\alias{keyword_count_top}
\title{Top keyword counts}
\usage{
keyword_count_top(wos, top_n = 5)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{top_n:}{the number of keywords}
}
\description{
The keyword counts are grouped by publication year. When keywords are tied in their raking, they are all returned.
}
\examples{
keyword_count_top()
}
\keyword{keyword}
