% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.r
\name{keyword_count_top_plot}
\alias{keyword_count_top_plot}
\title{Plot top keyword counts}
\usage{
keyword_count_top_plot(wos, top_n = 5)
}
\arguments{
\item{wos_key_count:}{the dataframe with keyword counts}

\item{top_n:}{the number of keywords}
}
\description{
Plot top keyword counts
}
\examples{
keyword_count_top_plot()
}
\keyword{keyword}
