% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{keywords}
\alias{keywords}
\title{keywords from WoS DE field separated}
\usage{
keywords(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
The function filters only for journal articles
}
\examples{
keywords()
}
\keyword{keyword}
