% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_data}
\alias{load_data}
\title{Load tsv file from Web of Science}
\usage{
load_data(file)
}
\arguments{
\item{file:}{the tsv of WoS references}
}
\description{
This function assumes that all the files have been joined, and all quote marks removed
}
\examples{
load_data()
}
\keyword{tsv}
