% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_directory}
\alias{load_directory}
\title{Load directory of WoS records}
\usage{
load_directory(dir)
}
\arguments{
\item{dir:}{the directory of files}
}
\description{
This function assumes that all the files have been zipped and all quote marks removed
}
\examples{
load_zip()
}
\keyword{WoS}
\keyword{directory,}
\keyword{files}
