% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_zip}
\alias{load_zip}
\title{Load zipfile of WoS records}
\usage{
load_zip(zipfile)
}
\arguments{
\item{zipfile:}{the collection of files}
}
\description{
This function assumes that all the files have been zipped and all quote marks removed
}
\examples{
load_zip()
}
\keyword{zip}
