% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_over_time.r
\name{peak_terms_monthly}
\alias{peak_terms_monthly}
\title{Plot  terms from TI, DE or some other field that have peakdd
over time.}
\usage{
peak_terms_monthly(wos, field = "TI", combine = FALSE, terms_to_plot = 16)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{field:}{individual field to use}

\item{combine:}{combine all relevant fields}

\item{terms_to_plot:}{the top or lowest n terms}
}
\description{
Plot  terms from TI, DE or some other field that have peakdd
over time.
}
\examples{
peak_terms_monthly()
}
\keyword{time}
