% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.r
\name{plot_keyword_count_top}
\alias{plot_keyword_count_top}
\title{Plot top keyword counts}
\usage{
plot_keyword_count_top(wos_key_count)
}
\arguments{
\item{wos_key_count:}{the dataframe with keyword counts}
}
\description{
Plot top keyword counts
}
\examples{
plot_keyword_count_top()
}
\keyword{keyword}
