% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{search_term}
\alias{search_term}
\title{Search title, abstract and keywords for terms}
\usage{
search_term(wos, terms, view = FALSE, format_as_bibtex = FALSE)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{term:}{the terms to search for as a vector; this can be a regex}

\item{view:}{show in Viewer}

\item{format_as_bibtex:}{return a single BibTex citation string [@A_2010; @B_2012]}
}
\description{
The search is run across all relevant fields
}
\examples{
search_term()
}
\keyword{search}
