% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{search_title_abstract_keywords}
\alias{search_title_abstract_keywords}
\title{Search title, abstract and keywords for terms}
\usage{
search_title_abstract_keywords(wos, terms, view = FALSE)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{term:}{the terms to search for a vector; this can be a regex}

\item{view:}{show in Viewer}
}
\description{
The search is run across all relevant fields
}
\examples{
search_title_abstract_keywords(wos, terms)
}
\keyword{search}
