% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_graph.r
\name{term_graph}
\alias{term_graph}
\title{Graph term pairs}
\usage{
term_graph(wos, min_n = 20, field = "TI")
}
\arguments{
\item{min_n}{the minimum number of pairs to include in the graph}

\item{field}{the Web of Science field to use - TI, AB or DE (title, abstract or keyword)}

\item{wos:}{the dataframe of references}
}
\description{
Graph term pairs
}
\examples{
term_graph()
}
\keyword{terms}
