% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_over_time.r
\name{terms_over_time}
\alias{terms_over_time}
\title{Plot occurrence of terms from TI, DE or some other field
over time.}
\usage{
terms_over_time(wos, field = "TI", combine = TRUE, plt = TRUE,
  terms_to_plot = 16)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{field:}{individual field to use}

\item{combine:}{combine all relevant fields}

\item{plt:}{whether to plot of not}

\item{terms_to_plot:}{the top or lowest n terms}
}
\description{
Plot occurrence of terms from TI, DE or some other field
over time.
}
\examples{
terms_over_time()
}
\keyword{time}
