% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.r
\name{tfidf_plot}
\alias{tfidf_plot}
\title{Plot Term Frequency-Inverse document frequency  of all words}
\usage{
tfidf_plot(wos_words, terms = 20)
}
\arguments{
\item{wos:_words:}{the dataframe of word counts for each document}

\item{terms:}{the number of terms to plot}
}
\description{
Plot Term Frequency-Inverse document frequency  of all words
}
\examples{
tfidf_plot()
}
\keyword{tfidf}
