% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{title_bigrams}
\alias{title_bigrams}
\title{Count of title bigrams}
\usage{
title_bigrams(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Count of title bigrams
}
\examples{
title_bigrams()
}
\keyword{title}
