% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{title_bigrams_count}
\alias{title_bigrams_count}
\title{Title bigrams sorted by frequency}
\usage{
title_bigrams_count(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Title bigrams sorted by frequency
}
\examples{
title_bigrams_count()
}
\keyword{bigram}
\keyword{title;}
