% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{title_bigrams_filtered}
\alias{title_bigrams_filtered}
\title{Title bigrams separated and filtered for stopwords}
\usage{
title_bigrams_filtered(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Title bigrams separated and filtered for stopwords
}
\examples{
title_bigrams_filtered()
}
\keyword{title}
