% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.r
\name{title_bigrams_separated}
\alias{title_bigrams_separated}
\title{Title bigrams separated into two columns}
\usage{
title_bigrams_separated(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Title bigrams separated into two columns
}
\examples{
title_bigrams_separated()
}
\keyword{title}
