% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.r
\name{title_word_count}
\alias{title_word_count}
\title{Retrieve count of title words}
\usage{
title_word_count(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Retrieve count of title words
}
\examples{
title_word_count()
}
\keyword{authors}
