% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{top_authors}
\alias{top_authors}
\title{Display top n first authors}
\usage{
top_authors(wos, n = 30)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{n:}{the number of authors to return}
}
\description{
Display top n first authors
}
\examples{
top_authors()
}
\keyword{authors}
