% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.r
\name{words_all}
\alias{words_all}
\title{Frequencies of all words in abstracts}
\usage{
words_all(wos, field = "AB")
}
\arguments{
\item{wos:}{the dataframe of references}

\item{field:}{the field to count}
}
\description{
Frequencies of all words in abstracts
}
\examples{
words_all()
}
\keyword{words}
