% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.r
\name{words_all_ranked_frequencies}
\alias{words_all_ranked_frequencies}
\title{Ranked frequencies of all words in abstracts}
\usage{
words_all_ranked_frequencies(wos, field = "AB", plot = FALSE)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{field:}{the field to count}

\item{plot:}{whether to plot the ranked frequencies}
}
\description{
Ranked frequencies of all words in abstracts
}
\examples{
words_all_ranked_frequencies()
}
\keyword{words}
