% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{year_count}
\alias{year_count}
\title{Number of publications by year}
\usage{
year_count(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Number of publications by year
}
\examples{
year_count()
}
\keyword{years}
