% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{cited_reference_count}
\alias{cited_reference_count}
\title{Counted cited references}
\usage{
cited_reference_count(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Counted cited references
}
\examples{
cited_reference_count()
}
\keyword{citations}
