% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{cited_reference_count}
\alias{cited_reference_count}
\title{Counted cited references
Returns the n top cited references grouped by year}
\usage{
cited_reference_count(wos, n = 20)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{n:}{the number to return}
}
\description{
Counted cited references
Returns the n top cited references grouped by year
}
\examples{
cited_reference_count()
}
\keyword{citations}
