% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{cited_references_gather}
\alias{cited_references_gather}
\title{Aggregate cited references}
\usage{
cited_references_gather(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Aggregate cited references
}
\examples{
cited_references_gather()
}
\keyword{citations}
