% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.r
\name{field_tfidf}
\alias{field_tfidf}
\title{Count term frequency/inverse document frequencies
TODO: implement for any field; at the moment it works for abstracts}
\usage{
field_tfidf(wos, field, extra_stopwords = "")
}
\arguments{
\item{wos:}{the dataframe of references}

\item{field:}{field to construct tfidf matrix for}

\item{extra_stopwords:}{any extra stopwords to exclude}
}
\description{
Count term frequency/inverse document frequencies
TODO: implement for any field; at the moment it works for abstracts
}
\examples{
field_tfidf()
}
\keyword{tfidf}
