% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.r
\name{keyword_count}
\alias{keyword_count}
\title{Count of top keywords}
\usage{
keyword_count(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Count of top keywords
}
\examples{
keyword_count()
}
\keyword{keyword}
