% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.r
\name{keyword_count_top}
\alias{keyword_count_top}
\title{Top keyword counts}
\usage{
keyword_count_top(wos, n = 5)
}
\arguments{
\item{wos:}{the dataframe of references}

\item{n:}{the number of keywords}
}
\description{
Top keyword counts
}
\examples{
keyword_count_top()
}
\keyword{keyword}
