% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_format.r
\name{latex_format}
\alias{latex_format}
\title{Return any set of WoS references as latex reference}
\usage{
latex_format(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
e.g. [@Derrida_2001a].
TODO: function that deals with van/von/de and other double barrel names 
TODO: multiple references for the same year.
}
\examples{
latex_format(wos)
}
