% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{title_words}
\alias{title_words}
\title{Retrieve count of title words}
\usage{
title_words(wos)
}
\arguments{
\item{wos:}{the dataframe of references}
}
\description{
Retrieve count of title words
}
\examples{
title_words()
}
\keyword{authors}
